implement fozztexx changes correct turnaround wait for ATN release after UNTALK replace timeoutWait() in some places

+++ lib/meatloaf/archive/lbr.cpp
@@ -129,9 +129,9 @@ bool LBRMStream::seekEntry(uint16_t index)
     return true;
 }
 
-uint16_t LBRMStream::readFile(uint8_t *buf, uint16_t size)
+uint32_t LBRMStream::readFile(uint8_t *buf, uint32_t size)
 {
-    uint16_t bytesRead = 0;
+    uint32_t bytesRead = 0;
 
     if (_position < 2)
     {
