+++ lib/meatloaf/disk/d64.h
@@ -242,7 +242,7 @@ public:
     }
 
     virtual bool seekPath(std::string path) override;
-    uint16_t readFile(uint8_t* buf, uint16_t size) override;
+    uint32_t readFile(uint8_t* buf, uint32_t size) override;
 
     Header header;      // Directory header data
     Entry entry;        // Directory entry data
