+++ lib/meatloaf/tape/tap.cpp
@@ -67,8 +67,8 @@ bool TAPMStream::seekEntry( uint16_t index )
 }
 
 
-uint16_t TAPMStream::readFile(uint8_t* buf, uint16_t size) {
-    uint16_t bytesRead = 0;
+uint32_t TAPMStream::readFile(uint8_t* buf, uint32_t size) {
+    uint32_t bytesRead = 0;
 
     bytesRead += containerStream->read(buf, size);
     _position += bytesRead;
